Moved the Zone class to it's own file
