+++ src/main/java/org/virtue/game/entity/combat/impl/AttackHandler.java
@@ -3,7 +3,7 @@ package org.virtue.game.entity.combat.impl;
 import org.virtue.game.entity.Entity;
 import org.virtue.game.entity.combat.AttackInfo;
 import org.virtue.game.entity.combat.CombatStyle;
-import org.virtue.game.map.zone.Projectile;
+import org.virtue.game.map.prot.Projectile;
 import org.virtue.network.protocol.update.block.SpotAnimationBlock;
 import org.virtue.utility.RandomExt;
