+++ src/main/java/org/virtue/game/entity/combat/impl/dragonfire/DragonfireHandler.java
@@ -5,7 +5,7 @@ import org.virtue.game.entity.combat.AttackInfo;
 import org.virtue.game.entity.combat.CombatStyle;
 import org.virtue.game.entity.combat.impl.AttackHandler;
 import org.virtue.game.entity.combat.impl.CombatUtils;
-import org.virtue.game.map.zone.Projectile;
+import org.virtue.game.map.prot.Projectile;
 
 /**
  * Handles dragonfire attackss.
