+++ src/main/java/org/virtue/game/entity/combat/impl/range/RangeAttackHandler.java
@@ -8,7 +8,7 @@ import org.virtue.game.entity.combat.impl.CombatUtils;
 import org.virtue.game.entity.combat.impl.formula.Formula;
 import org.virtue.game.entity.player.Player;
 import org.virtue.game.entity.player.inv.Item;
-import org.virtue.game.map.zone.Projectile;
+import org.virtue.game.map.prot.Projectile;
 
 /**
  * Handles range attacks.
