+++ src/main/java/org/virtue/game/entity/npc/script/ArcherCombat.java
@@ -9,7 +9,7 @@ import org.virtue.game.entity.combat.impl.AttackHandler;
 import org.virtue.game.entity.combat.impl.FollowingType;
 import org.virtue.game.entity.combat.impl.range.RangeAttackEvent;
 import org.virtue.game.entity.combat.impl.range.RangeAttackHandler;
-import org.virtue.game.map.zone.Projectile;
+import org.virtue.game.map.prot.Projectile;
 import org.virtue.network.protocol.update.block.AnimationBlock;
 import org.virtue.network.protocol.update.block.SpotAnimationBlock;
