+++ src/main/java/org/virtue/game/entity/player/Player.java
@@ -70,11 +70,11 @@ import org.virtue.game.entity.player.var.VarKey;
 import org.virtue.game.entity.player.var.VarRepository;
 import org.virtue.game.entity.player.widget.WidgetManager;
 import org.virtue.game.map.GroundItem;
+import org.virtue.game.map.prot.Projectile;
+import org.virtue.game.map.prot.ZoneUpdatePacket;
 import org.virtue.game.map.CoordGrid;
 import org.virtue.game.map.square.DynamicMapSquare;
 import org.virtue.game.map.square.MapSquare;
-import org.virtue.game.map.zone.Projectile;
-import org.virtue.game.map.zone.ZoneUpdatePacket;
 import org.virtue.game.parser.ParserType;
 import org.virtue.network.event.GameEventDispatcher;
 import org.virtue.network.event.context.impl.out.ZoneUpdateEventContext;
