+++ src/main/java/org/virtue/game/map/prot/DeleteObject.java
@@ -19,10 +19,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.virtue.game.map.zone;
+package org.virtue.game.map.prot;
 
 import org.virtue.game.entity.Entity;
-import org.virtue.game.map.SceneLocation;
+import org.virtue.game.map.GroundItem;
 import org.virtue.network.event.buffer.OutboundBuffer;
 import org.virtue.game.map.CoordGrid;
 
@@ -31,16 +31,14 @@ import org.virtue.game.map.CoordGrid;
  * @author Frosty Teh Snowman <skype:travis.mccorkle>
  * @author Arthur <skype:arthur.behesnilian>
  * @author Sundays211
- * @since 4/11/2014
+ * @since 2/11/2014
  */
-public class DeleteLocation implements ZoneUpdatePacket {
+public class DeleteObject implements ZoneUpdatePacket {
 	
-	private int flags;
-	private CoordGrid coord;
+	private GroundItem object;
 	
-	public DeleteLocation (SceneLocation object) {
-		this.flags = (object.getRotation() & 0x3) | (object.getShape().getId() << 2);
-		this.coord = object.getTile();
+	public DeleteObject (GroundItem object) {
+		this.object = object;
 	}
 
 	/* (non-Javadoc)
@@ -48,7 +46,7 @@ public class DeleteLocation implements ZoneUpdatePacket {
 	 */
 	@Override
 	public ZoneProtocol getType() {
-		return ZoneProtocol.LOC_DEL;
+		return ZoneProtocol.OBJ_DEL;
 	}
 
 	/* (non-Javadoc)
@@ -56,8 +54,8 @@ public class DeleteLocation implements ZoneUpdatePacket {
 	 */
 	@Override
 	public void encode(OutboundBuffer buffer, Entity player) {
-		buffer.putByte(((coord.getX() % 8) & 0x7) << 4 | (coord.getY() % 8) & 0x7);
-		buffer.putS(flags);
+		buffer.putLEShort(object.getId());
+		buffer.putC((object.getOffsetX() & 0x7) << 4 | object.getOffsetY() & 0x7);
 	}
 
 	/* (non-Javadoc)
@@ -65,7 +63,7 @@ public class DeleteLocation implements ZoneUpdatePacket {
 	 */
 	@Override
 	public CoordGrid getTile() {
-		return coord;
+		return object.getTile();
 	}
 
 }
